requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Model
timm>=0.9.0               # EfficientNet-B4 pretrained
onnx>=1.14.0
onnxruntime>=1.16.0

# Explainability
shap>=0.42.0
grad-cam>=1.4.8           # pytorch-grad-cam

# Federated learning
# Using custom lightweight implementation (no external FL framework needed)

# Visualization & logging
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0
tqdm>=4.65.0

# Utilities
PyYAML>=6.0
